pyproject.toml
[tool.poetry]
name = "choose_one_be"
version = "0.1.0"
description = ""
authors = ["Yurii <jurakulek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.12.1"
fastapi = "^0.61.1"
PyJWT = "^1.7.1"
redis = "^3.5.3"
Werkzeug = "^1.0.1"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"
pytest-dotenv = "^0.5.2"
celery = "^5.0.0"
environs = "^8.0.0"
pyhumps = "^1.6.1"
pydantic = {extras = ["email"], version = "^1.6.1"}
Jinja2 = "^2.11.2"
mailjet-rest = "^1.3.3"
loguru = "^0.5.3"
asyncpg = "^0.21.0"
pytest-mock = "^3.3.1"
python-slugify = "^4.0.1"
alembic = "^1.4.3"
SQLAlchemy-Utils = "^0.36.8"
psycopg2-binary = "^2.8.6"
flake8 = "^3.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
